release: v0.15.0 (#83)
* feat: remove a file (#82)

* feat: remove endpoint

Endpoint to remove file completed, in addition to adding tests

* docs(http): Update bruno collection

---------

+++ src/controllers/files/remove_file/_handler.py
@@ -1,8 +1,12 @@
 from src.config.soap_client import soap_client
+from src.lib.helpers import is_valid_uuid
 
 
 def remove_file_handler(token, file_uuid):
     try:
+        if not is_valid_uuid(file_uuid):
+            return {"msg": "Not valid file UUID provided"}, 400
+
         request_data = {"fileUUID": file_uuid, "token": token}
         response = soap_client.service.file_delete(request_data)
